pyproject.toml
[tool.poetry]
name = "parkingLot"
version = "0.1.0"
description = ""
authors = ["Matvey <matveykawork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-split-settings = "^1.2.0"
gunicorn = "^21.2.0"
django-debug-toolbar = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
